requirements.txt
# Core ML/DL
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Data fetching
requests>=2.31.0
python-dotenv>=1.0.0

# Scheduling & caching
apscheduler>=3.10.0
redis>=5.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
joblib>=1.3.0
tqdm>=4.66.0
loguru>=0.7.0
